# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Newsletter Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface VideoRecord {
  title: string;
  url: string;
  channel: string;
  /** @nullable */
  transcript: string | null;
  /** @nullable */
  transcript_error?: string | null;
}

export interface RunRecord {
  total_videos: number;
  email_sent: boolean;
  videos: VideoRecord[];
}

/**
 * Map of ISO date strings to arrays of RunRecord
 */
export interface RunsData {[key: string]: RunRecord[]}

export interface GmailStatus {
  date: string;
  /**
     * true if email found, false if not found, null if Gmail check failed
     * @nullable
     */
  found: boolean | null;
  /** true if Gmail credentials are not configured */
  unavailable: boolean;
}

export interface DashboardSummary {
  totalRuns: number;
  last30DaysRuns: number;
  last30DaysEmailsSent: number;
  /** Fraction of videos successfully transcribed in last 30 days (0-1) */
  last30DaysTranscriptRate: number;
  totalVideosProcessed: number;
}