# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Newsletter Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalRuns: number;
  last30DaysRuns: number;
  last30DaysEmailsSent: number;
  /** Fraction of videos successfully transcribed in last 30 days (0-1) */
  last30DaysTranscriptRate: number;
  totalVideosProcessed: number;
}