# lib/api-zod/src/generated/types/gmailStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Newsletter Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface GmailStatus {
  date: string;
  /**
     * true if email found, false if not found, null if Gmail check failed
     * @nullable
     */
  found: boolean | null;
  /** true if Gmail credentials are not configured */
  unavailable: boolean;
}