# lib/api-zod/src/generated/types/runRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Newsletter Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { VideoRecord } from './videoRecord';

export interface RunRecord {
  total_videos: number;
  email_sent: boolean;
  videos: VideoRecord[];
}