# lib/api-zod/src/generated/types/runsData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Newsletter Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { RunRecord } from './runRecord';

/**
 * Map of ISO date strings to arrays of RunRecord
 */
export interface RunsData {[key: string]: RunRecord[]}