# lib/api-zod/src/generated/types/videoRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Newsletter Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface VideoRecord {
  title: string;
  url: string;
  channel: string;
  /** @nullable */
  transcript: string | null;
  /** @nullable */
  transcript_error?: string | null;
}